fix(tests): increase acceptable runtime limit (#66)
On slower hardware, or on newer hardware that did not have the container
images fetched on the worker nodes yet, or on a system that was busy and
ran many DEMO examples in parallel, the actual workflow execution
finished successfully but was slower than on a usual idle and preloaded
system. This commit allows for more liberal runtime limits in order to
fit these slower testing scenarios.